rename gathered data file
