+++ depthai_nodes/message/gathered_data.py
